[improve]Improve the way dorisSystem create table
