+++ flink-doris-connector/src/main/java/org/apache/doris/flink/catalog/doris/DorisSystem.java
@@ -25,7 +25,6 @@ import org.apache.commons.compress.utils.Lists;
 import org.apache.doris.flink.cfg.DorisConnectionOptions;
 import org.apache.doris.flink.connection.JdbcConnectionProvider;
 import org.apache.doris.flink.connection.SimpleJdbcConnectionProvider;
-import org.apache.doris.flink.exception.CreateTableException;
 import org.apache.doris.flink.exception.DorisRuntimeException;
 import org.apache.doris.flink.exception.DorisSystemException;
 import org.slf4j.Logger;
@@ -41,7 +40,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 import static org.apache.flink.util.Preconditions.checkArgument;
 
@@ -141,81 +139,7 @@ public class DorisSystem implements Serializable {
     }
 
     public static String buildCreateTableDDL(TableSchema schema) {
-        StringBuilder sb = new StringBuilder("CREATE TABLE IF NOT EXISTS ");
-        sb.append(identifier(schema.getDatabase()))
-                .append(".")
-                .append(identifier(schema.getTable()))
-                .append("(");
-
-        Map<String, FieldSchema> fields = schema.getFields();
-        List<String> keys = schema.getKeys();
-        // append keys
-        for (String key : keys) {
-            if (!fields.containsKey(key)) {
-                throw new CreateTableException("key " + key + " not found in column list");
-            }
-            FieldSchema field = fields.get(key);
-            buildColumn(sb, field, true);
-        }
-
-        // append values
-        for (Map.Entry<String, FieldSchema> entry : fields.entrySet()) {
-            if (keys.contains(entry.getKey())) {
-                continue;
-            }
-            FieldSchema field = entry.getValue();
-            buildColumn(sb, field, false);
-        }
-        sb = sb.deleteCharAt(sb.length() - 1);
-        sb.append(" ) ");
-        // append uniq model
-        if (DataModel.UNIQUE.equals(schema.getModel())) {
-            sb.append(schema.getModel().name())
-                    .append(" KEY(")
-                    .append(String.join(",", identifier(schema.getKeys())))
-                    .append(")");
-        }
-
-        // append table comment
-        if (!StringUtils.isNullOrWhitespaceOnly(schema.getTableComment())) {
-            sb.append(" COMMENT '").append(quoteComment(schema.getTableComment())).append("' ");
-        }
-
-        // append distribute key
-        sb.append(" DISTRIBUTED BY HASH(")
-                .append(String.join(",", identifier(schema.getDistributeKeys())))
-                .append(")");
-
-        Map<String, String> properties = schema.getProperties();
-        if (schema.getTableBuckets() != null) {
-
-            int bucketsNum = schema.getTableBuckets();
-            if (bucketsNum <= 0) {
-                throw new CreateTableException("The number of buckets must be positive.");
-            }
-            sb.append(" BUCKETS ").append(bucketsNum);
-        } else {
-            sb.append(" BUCKETS AUTO ");
-        }
-        // append properties
-        int index = 0;
-        for (Map.Entry<String, String> entry : properties.entrySet()) {
-            if (index == 0) {
-                sb.append(" PROPERTIES (");
-            }
-            if (index > 0) {
-                sb.append(",");
-            }
-            sb.append(quoteProperties(entry.getKey()))
-                    .append("=")
-                    .append(quoteProperties(entry.getValue()));
-            index++;
-
-            if (index == (schema.getProperties().size())) {
-                sb.append(")");
-            }
-        }
-        return sb.toString();
+        return DorisSchemaFactory.generateCreateTableDDL(schema);
     }
 
     public Map<String, String> getTableFieldNames(String databaseName, String tableName) {
@@ -244,19 +168,6 @@ public class DorisSystem implements Serializable {
         }
     }
 
-    private static void buildColumn(StringBuilder sql, FieldSchema field, boolean isKey) {
-        String fieldType = field.getTypeString();
-        if (isKey && DorisType.STRING.equals(fieldType)) {
-            fieldType = String.format("%s(%s)", DorisType.VARCHAR, 65533);
-        }
-        sql.append(identifier(field.getName())).append(" ").append(fieldType);
-
-        if (field.getDefaultValue() != null) {
-            sql.append(" DEFAULT " + quoteDefaultValue(field.getDefaultValue()));
-        }
-        sql.append(" COMMENT '").append(quoteComment(field.getComment())).append("',");
-    }
-
     public static String quoteDefaultValue(String defaultValue) {
         // DEFAULT current_timestamp not need quote
         if (defaultValue.equalsIgnoreCase("current_timestamp")) {
@@ -273,10 +184,6 @@ public class DorisSystem implements Serializable {
         }
     }
 
-    private static List<String> identifier(List<String> name) {
-        return name.stream().map(DorisSystem::identifier).collect(Collectors.toList());
-    }
-
     public static String identifier(String name) {
         if (name.startsWith("`") && name.endsWith("`")) {
             return name;
@@ -288,9 +195,5 @@ public class DorisSystem implements Serializable {
         String[] dbTable = tableIdentifier.split("\\.");
         Preconditions.checkArgument(dbTable.length == 2);
         return identifier(dbTable[0]) + "." + identifier(dbTable[1]);
-    }
-
-    private static String quoteProperties(String name) {
-        return "'" + name + "'";
     }
 }
