+++ flink-doris-connector/src/test/java/org/apache/doris/flink/catalog/doris/DorisSchemaFactoryTest.java
@@ -46,6 +46,13 @@ public class DorisSchemaFactoryTest {
 
     @Test
     public void testCreateTableSchema() {
+        TableSchema tableSchema = buildCreateTableSchema();
+        Assert.assertEquals(
+                "TableSchema{database='doris', table='create_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=email, model=UNIQUE, distributeKeys=email, properties={light_schema_change=true}, tableBuckets=null}",
+                tableSchema.toString());
+    }
+
+    private TableSchema buildCreateTableSchema() {
         String dorisTable = "doris.create_tab";
         String[] dbTable = dorisTable.split("\\.");
         Preconditions.checkArgument(dbTable.length == 2);
@@ -58,21 +65,24 @@ public class DorisSchemaFactoryTest {
         List<String> pkKeys = Collections.singletonList("email");
         Map<String, String> tableProperties = new HashMap<>();
         String tableComment = "auto_tab_comment";
-        TableSchema tableSchema =
-                DorisSchemaFactory.createTableSchema(
-                        dbTable[0],
-                        dbTable[1],
-                        columnFields,
-                        pkKeys,
-                        new DorisTableConfig(tableProperties),
-                        tableComment);
-        Assert.assertEquals(
-                "TableSchema{database='doris', table='create_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=email, model=UNIQUE, distributeKeys=email, properties={light_schema_change=true}, tableBuckets=null}",
-                tableSchema.toString());
+        return DorisSchemaFactory.createTableSchema(
+                dbTable[0],
+                dbTable[1],
+                columnFields,
+                pkKeys,
+                new DorisTableConfig(tableProperties),
+                tableComment);
     }
 
     @Test
     public void testCreateTableSchemaTableBuckets() {
+        TableSchema tableSchema = buildCreateTableSchemaTableBuckets();
+        Assert.assertEquals(
+                "TableSchema{database='doris', table='create_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=email, model=UNIQUE, distributeKeys=email, properties={replication_num=2, light_schema_change=true}, tableBuckets=40}",
+                tableSchema.toString());
+    }
+
+    private TableSchema buildCreateTableSchemaTableBuckets() {
         String dorisTable = "doris.create_tab";
         String[] dbTable = dorisTable.split("\\.");
         Preconditions.checkArgument(dbTable.length == 2);
@@ -87,21 +97,24 @@ public class DorisSchemaFactoryTest {
         tableProperties.put("table-buckets", "create_tab:40, create_taba:10, tabs:12");
         tableProperties.put("replication_num", "2");
         String tableComment = "auto_tab_comment";
-        TableSchema tableSchema =
-                DorisSchemaFactory.createTableSchema(
-                        dbTable[0],
-                        dbTable[1],
-                        columnFields,
-                        pkKeys,
-                        new DorisTableConfig(tableProperties),
-                        tableComment);
-        Assert.assertEquals(
-                "TableSchema{database='doris', table='create_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=email, model=UNIQUE, distributeKeys=email, properties={replication_num=2, light_schema_change=true}, tableBuckets=40}",
-                tableSchema.toString());
+        return DorisSchemaFactory.createTableSchema(
+                dbTable[0],
+                dbTable[1],
+                columnFields,
+                pkKeys,
+                new DorisTableConfig(tableProperties),
+                tableComment);
     }
 
     @Test
     public void testCreateDuplicateTableSchema() {
+        TableSchema tableSchema = buildCreateDuplicateTableSchema();
+        Assert.assertEquals(
+                "TableSchema{database='doris', table='dup_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=name, model=DUPLICATE, distributeKeys=name, properties={replication_num=1, light_schema_change=true}, tableBuckets=null}",
+                tableSchema.toString());
+    }
+
+    private TableSchema buildCreateDuplicateTableSchema() {
         String dorisTable = "doris.dup_tab";
         String[] dbTable = dorisTable.split("\\.");
         Preconditions.checkArgument(dbTable.length == 2);
@@ -114,16 +127,39 @@ public class DorisSchemaFactoryTest {
         Map<String, String> tableProperties = new HashMap<>();
         tableProperties.put("replication_num", "1");
         String tableComment = "auto_tab_comment";
-        TableSchema tableSchema =
-                DorisSchemaFactory.createTableSchema(
-                        dbTable[0],
-                        dbTable[1],
-                        columnFields,
-                        new ArrayList<>(),
-                        new DorisTableConfig(tableProperties),
-                        tableComment);
+        return DorisSchemaFactory.createTableSchema(
+                dbTable[0],
+                dbTable[1],
+                columnFields,
+                new ArrayList<>(),
+                new DorisTableConfig(tableProperties),
+                tableComment);
+    }
+
+    @Test
+    public void testGenerateCreateTableDDL() {
+        TableSchema tableSchema = buildCreateTableSchema();
+        String ddl = DorisSchemaFactory.generateCreateTableDDL(tableSchema);
         Assert.assertEquals(
-                "TableSchema{database='doris', table='dup_tab', tableComment='auto_tab_comment', fields={name=FieldSchema{name='name', typeString='VARVHAR(100)', defaultValue='null', comment='Name_test'}, id=FieldSchema{name='id', typeString='INT', defaultValue='100', comment='int_test'}, age=FieldSchema{name='age', typeString='INT', defaultValue='null', comment='null'}, email=FieldSchema{name='email', typeString='VARCHAR(100)', defaultValue='email@doris.com', comment='e'}}, keys=name, model=DUPLICATE, distributeKeys=name, properties={replication_num=1, light_schema_change=true}, tableBuckets=null}",
-                tableSchema.toString());
+                "CREATE TABLE IF NOT EXISTS `doris`.`create_tab`(`email` VARCHAR(100) DEFAULT 'email@doris.com' COMMENT 'e',`name` VARVHAR(100) COMMENT 'Name_test',`id` INT DEFAULT '100' COMMENT 'int_test',`age` INT COMMENT '' ) UNIQUE KEY(`email`) COMMENT 'auto_tab_comment'  DISTRIBUTED BY HASH(`email`) BUCKETS AUTO  PROPERTIES ('light_schema_change'='true')",
+                ddl);
+    }
+
+    @Test
+    public void testGenerateCreateTableDDLBuckets() {
+        TableSchema tableSchema = buildCreateTableSchemaTableBuckets();
+        String ddl = DorisSchemaFactory.generateCreateTableDDL(tableSchema);
+        Assert.assertEquals(
+                "CREATE TABLE IF NOT EXISTS `doris`.`create_tab`(`email` VARCHAR(100) DEFAULT 'email@doris.com' COMMENT 'e',`name` VARVHAR(100) COMMENT 'Name_test',`id` INT DEFAULT '100' COMMENT 'int_test',`age` INT COMMENT '' ) UNIQUE KEY(`email`) COMMENT 'auto_tab_comment'  DISTRIBUTED BY HASH(`email`) BUCKETS 40 PROPERTIES ('replication_num'='2','light_schema_change'='true')",
+                ddl);
+    }
+
+    @Test
+    public void testGenerateCreateTableDDLDuplicate() {
+        TableSchema tableSchema = buildCreateDuplicateTableSchema();
+        String ddl = DorisSchemaFactory.generateCreateTableDDL(tableSchema);
+        Assert.assertEquals(
+                "CREATE TABLE IF NOT EXISTS `doris`.`dup_tab`(`name` VARVHAR(100) COMMENT 'Name_test',`id` INT DEFAULT '100' COMMENT 'int_test',`age` INT COMMENT '',`email` VARCHAR(100) DEFAULT 'email@doris.com' COMMENT 'e' )  COMMENT 'auto_tab_comment'  DISTRIBUTED BY HASH(`name`) BUCKETS AUTO  PROPERTIES ('replication_num'='1','light_schema_change'='true')",
+                ddl);
     }
 }
